requirements.txt
viam-sdk
pyinstaller
smbus2